refactor to reduce duplication for media model
